+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### Next release
 - Changed Jenkins logo on monitor
 - Improved rebuild plugin integration
+- Analytics tracking
 
 #### v0.4.0-build.1565952664 ~ 29.08.16
 - Job health status indication